Handle edge case where no browser exists
In case the running machine do not have any browser installed,
the wmic command will report on stderr with
"No Instance(s) Available.". We don't handle this case as an
actual exception, so we do not throw. The script will never reach
fkill.

+++ windows.js
@@ -26,7 +26,7 @@ export default async function killTabs(options) {
 	const args = ['process', 'where', browsers.join(' or '), 'get', 'CommandLine,ProcessId', '/format:list'];
 	const response = await promisify(childProcess.execFile)('wmic', args, {maxBuffer: TEN_MEBIBYTE});
 
-	if (response.stderr) {
+	if (response.stderr && !response.stderr.includes('No Instance(s) Available.')) {
 		throw new Error('Failed to kill tabs.', {cause: new Error(response.stderr)});
 	}
